requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
celery[redis]>=5.3.4
redis>=5.0.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.9
qdrant-client>=1.7.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
google-generativeai>=0.8.0
google-genai>=0.8.0
langchain>=0.1.0
langchain-community>=0.0.10
langchain-text-splitters>=0.0.1
beautifulsoup4>=4.12.0
requests>=2.31.0
lxml>=4.9.0
asyncpg>=0.28.0